+++ sklearn/model_selection/__init__.py
@@ -1,0 +1,28 @@
+from .partition import LeaveOneOut, LeavePOut, LeaveOneLabelOut, LeavePLabelOut
+from .partition import StratifiedKFold, StratifiedShuffleSplit, ShuffleSplit
+from .partition import KFold, check_cv, train_test_split
+from .validate import cross_val_score, permutation_test_score
+from .validate import learning_curve, validation_curve
+from .search import GridSearchCV, RandomizedSearchCV
+from .utils import ParameterGrid, ParameterSampler, fit_grid_point
+
+__all__ = ['KFold',
+           'LeaveOneLabelOut',
+           'LeaveOneOut',
+           'LeavePLabelOut',
+           'LeavePOut',
+           'ShuffleSplit',
+           'StratifiedKFold',
+           'StratifiedShuffleSplit',
+           'check_cv',
+           'cross_val_score',
+           'permutation_test_score',
+           'train_test_split',
+           'learning_curve',
+           'validation_curve',
+           'GridSearchCV',
+           'RandomizedSearchCV',
+           'ParameterGrid',
+           'ParameterSampler',
+           'fit_grid_point'
+           ]
